When possible, read in chunks when using readn
LambdaMemory has a cache now, which makes for more
solver-friendly constraints.

However, this cache is not used when we are evaluating
a `readn` of more than one word. For example, when
calculating a mapping, you need to `keccak(val . slot)`.
This change makes it so that we can make use of the
cache for `val` and for `slot` -- instead of reading
them via array-ops (which are slow).